refactor: renamed template folder
