+++ src/pages/Users/index.tsx
@@ -1,4 +1,4 @@
-import { UsersList } from '../../Templates/UsersList'
+import { UsersList } from '../../templates/UsersList'
 const Users: React.FC = () => {
   document.title = 'Uni.Cloud | Usuários'
   return <UsersList />
