+++ src/templates/AuthRegister/LoadingValidation.tsx
