+++ src/templates/AuthRegister/index.tsx
